Fixes for Xcode build

+++ examples/pxScene2d/external/build.sh
@@ -36,7 +36,6 @@ while (( "$#" )); do
   esac
 done
 
-#--------- CURL
 
 make_parallel=3
 
@@ -54,23 +53,33 @@ fi
 
 #--------- OPENSSL
 
-cd ${OPENSSL_DIR}
-if [ "$(uname)" != "Darwin" ]
+if [ ! -e ${OPENSSL_DIR}/lib/libcrypto.dylib ] ||
+   [ "$(uname)" != "Darwin" ]
 then
-./config -shared  --prefix=`pwd`
-else
-./Configure darwin64-x86_64-cc -shared --prefix=`pwd`
+
+  cd ${OPENSSL_DIR}
+
+  if [ "$(uname)" != "Darwin" ]
+  then
+    ./config -shared  --prefix=`pwd`
+  else
+    ./Configure darwin64-x86_64-cc -shared --prefix=`pwd`
+  fi
+
+  make clean
+  make "-j${make_parallel}"
+  make install -i
+
+  rm -rf libcrypto.a
+  rm -rf libssl.a
+  rm -rf lib/libcrypto.a
+  rm -rf lib/libssl.a
+
+  cd ..
+
+  export LD_LIBRARY_PATH="${OPENSSL_DIR}/:$LD_LIBRARY_PATH"
+  export DYLD_LIBRARY_PATH="${OPENSSL_DIR}/:$DYLD_LIBRARY_PATH"
 fi
-make clean
-make "-j${make_parallel}"
-make install -i
-rm -rf libcrypto.a
-rm -rf libssl.a
-rm -rf lib/libcrypto.a
-rm -rf lib/libssl.a
-cd ..
-export LD_LIBRARY_PATH="${OPENSSL_DIR}/:$LD_LIBRARY_PATH"
-export DYLD_LIBRARY_PATH="${OPENSSL_DIR}/:$DYLD_LIBRARY_PATH"
 
 #--------- PNG
 
@@ -82,8 +91,10 @@ then
 
   cd png
   ./configure --prefix=$EXT_INSTALL_PATH
+
   make all "-j${make_parallel}"
   make install
+
   # git update-index --assume-unchanged .      # ... help GIT out
   # git update-index --assume-unchanged src    # ... help GIT out
 
@@ -109,129 +120,138 @@ then
 fi
 #---------
 
-if [[ $# -eq 1 ]] && [[ $1 == "SPARK_ENABLE_VIDEO" ]]; then
-#--------graphite2
+#-------- GRAPHITE2
 
-if [ ! -e $EXT_INSTALL_PATH/lib/libgraphite2.la ]
+if [[ $# -eq 1 ]] && [[ $1 == "SPARK_ENABLE_VIDEO" ]]
 then
-  banner "graphite2"
 
-  ./graphite2/build.sh
+  if [ ! -e $EXT_INSTALL_PATH/lib/libgraphite2.la ]
+  then
+    banner "GRAPHITE2"
+
+    ./graphite2/build.sh
+  fi
+
 fi
 
 #--------
 
-#-------- pcre
+#-------- PCRE
 
 if [ ! -e $EXT_INSTALL_PATH/lib/libpcre.la ]
 then
-  banner "pcre"
+  banner "PCRE"
 
   ./pcre/build.sh
 fi
 
 #--------
 
-#--------icu
+#-------- ICU
 
 if [ ! -e $EXT_INSTALL_PATH/lib/libicudata.$LIBEXTN ]
 then
-  banner "icu"
+  banner "ICU"
 
   ./icu/build.sh
 fi
 
 #--------
 
-#-------- libffi
+#-------- LIBFFI
 
 if [ ! -e $EXT_INSTALL_PATH/lib/libffi.la ]
 then
-  banner "libffi"
+  banner "LIBFFI"
 
   ./libffi/build.sh
 fi
 
 #--------
 
-#--------gettext
+#-------- GETTEXT
 
 if [ ! -e $EXT_INSTALL_PATH/lib/libintl.la ]
 then
-  banner "gettext"
+  banner "GETTEXT"
 
   ./gettext/build.sh
 fi
 
 #--------
 
-#--------glib
+#-------- GLIB
 
 if [ ! -e $EXT_INSTALL_PATH/lib/libglib-2.0.la ]
 then
-  banner "glib"
+  banner "GLIB"
 
   ./glib/build.sh
 fi
 
 #--------
-fi
 
 #--------- GIF
 
-banner "GIF"
-
-cd gif
-
-[ -d patches ] || mkdir -p patches
-
-if [ "$(uname)" == "Darwin" ]; then
-  [ -d patches/series ] || echo 'giflib-5.1.9-mac.patch' >patches/series
-  cp ../giflib-5.1.9-mac.patch patches/
-else
-  [ -d patches/series ] || echo 'giflib-5.1.9.patch' >patches/series
-  cp ../giflib-5.1.9.patch patches/
-fi
-
-
-if [[ "$#" -eq "1" && "$1" == "--clean" ]]; then
-	quilt pop -afq || test $? = 2
-	rm -rf .libs/*
-elif [[ "$#" -eq "1" && "$1" == "--force-clean" ]]; then
-	git clean -fdx .
-	git checkout .
-	rm -rf .libs/*
-else
-	quilt push -aq || test $? = 2
-fi
-
-if [ ! -e ./.libs/libgif.7.dylib ] ||
-[ "$(uname)" != "Darwin" ]
+if [ ! -e ./gif/.libs/libgif.7.dylib ] ||
+   [ "$(uname)" != "Darwin" ]
 then
 
-  make "-j${make_parallel}"
+  banner "GIF"
 
-  [ -d .libs ] || mkdir -p .libs
-  if [ -e libgif.7.dylib ]
-  then
-      cp libgif.7.dylib .libs/libgif.7.dylib
-      cp libutil.7.dylib .libs/libutil.7.dylib
+  cd gif
 
-  elif [ -e libgif.so ]
-  then
-      cp libgif.so libgif.so.7
-      cp libutil.so libutil.so.7
-      cp libgif.so .libs/libgif.so
-      cp libutil.so .libs/libutil.so
-      cp libgif.so .libs/libgif.so.7
-      cp libutil.so .libs/libutil.so.7
+  [ -d patches ] || mkdir -p patches
+
+  if [ "$(uname)" == "Darwin" ]; then
+    [ -d patches/series ] || echo 'giflib-5.1.9-mac.patch' >patches/series
+    cp ../giflib-5.1.9-mac.patch patches/
+  else
+    [ -d patches/series ] || echo 'giflib-5.1.9.patch' >patches/series
+    cp ../giflib-5.1.9.patch patches/
   fi
+
+
+  if [[ "$#" -eq "1" && "$1" == "--clean" ]]; then
+    quilt pop -afq || test $? = 2
+    rm -rf .libs/*
+  elif [[ "$#" -eq "1" && "$1" == "--force-clean" ]]; then
+    git clean -fdx .
+    git checkout .
+    rm -rf .libs/*
+  else
+    quilt push -aq || test $? = 2
+  fi
+
+  if [ ! -e ./.libs/libgif.7.dylib ] ||
+  [ "$(uname)" != "Darwin" ]
+  then
+
+    make "-j${make_parallel}"
+
+    [ -d .libs ] || mkdir -p .libs
+    if [ -e libgif.7.dylib ]
+    then
+        cp libgif.7.dylib .libs/libgif.7.dylib
+        cp libutil.7.dylib .libs/libutil.7.dylib
+
+    elif [ -e libgif.so ]
+    then
+        cp libgif.so libgif.so.7
+        cp libutil.so libutil.so.7
+        cp libgif.so .libs/libgif.so
+        cp libutil.so .libs/libutil.so
+        cp libgif.so .libs/libgif.so.7
+        cp libutil.so .libs/libutil.so.7
+    fi
+  fi
+
+  git update-index --assume-unchanged dgif_lib.c     # ... help GIT out
+  git update-index --assume-unchanged Makefile       # ... help GIT out
+
+  cd ..
+
 fi
-
-git update-index --assume-unchanged dgif_lib.c     # ... help GIT out
-git update-index --assume-unchanged Makefile       # ... help GIT out
-
-cd ..
 
 #--------- FT
 
@@ -252,27 +272,28 @@ fi
 #---------
 
 if [[ $# -eq 1 ]] && [[ $1 == "SPARK_ENABLE_VIDEO" ]]; then
-#--------Fontconfig
 
-if [ ! -e $EXT_INSTALL_PATH/lib/libfontconfig.$LIBEXTN ]
-then
-  banner "Fontconfig"
+#-------- FONTCONFIG
 
-  ./fontconfig/build.sh
-fi
+  if [ ! -e $EXT_INSTALL_PATH/lib/libfontconfig.$LIBEXTN ]
+  then
+    banner "FONTCONFIG"
+
+    ./fontconfig/build.sh
+  fi
 
 #--------
 
-#--------harfbuzz
+#-------- HARFBUZZ
 
 if [ ! -e $EXT_INSTALL_PATH/lib/libharfbuzz.la ]
 then
-  banner "harfbuzz"
+  banner "HARFBUZZ"
 
   ./harfbuzz/build.sh
 fi
 #---------
-fi
+fi # SPARK_ENABLE_VIDEO
 
 #-------- openssl
 
@@ -300,7 +321,6 @@ if [ ! -e $EXT_INSTALL_PATH/lib/libcurl.la ]; then
     sed -i '' '/#define HAVE_CLOCK_GETTIME_MONOTONIC 1/d' lib/curl_config.h
   fi
 
-  
   make all "-j${make_parallel}"
   make install
   cd ..
@@ -369,7 +389,8 @@ if [ ! -e "libnode-v${NODE_VER}/libnode.dylib" ] ||
    [ "$(uname)" != "Darwin" ]
 then
 
-  banner "NODE"
+  banner "LIBNODE"
+
   if [ -e "node-v${NODE_VER}_mods.patch" ]
   then
     git apply "node-v${NODE_VER}_mods.patch"
@@ -434,10 +455,15 @@ fi
 
 #-------- NANOSVG
 
+if [ ! -e nanosvg/PATCHED ]
+then
 banner "NANOSVG"
 
-./nanosvg/build.sh
-git update-index --assume-unchanged ./nanosvg/src/nanosvgrast.h    # ... help GIT out
+  ./nanosvg/build.sh
+
+  git update-index --assume-unchanged ./nanosvg/src/nanosvgrast.h    # ... help GIT out
+fi
+
 #---------
 
 #-------- DUKTAPE
@@ -459,14 +485,21 @@ then
 
 fi
 
-#-------- spark-webgl
-export NODE_PATH=$NODE_PATH:`pwd`/../src/node_modules
-export PATH=`pwd`/node/deps/npm/bin/node-gyp-bin/:`pwd`/node/out/Release:$PATH
-cd spark-webgl
-node-gyp rebuild
-cd ..
+#-------- SPARK-WEBGL
 
-#-------- 
+if [ ! -d "spark-webgl/build" ]; then
+
+  export NODE_PATH=$NODE_PATH:`pwd`/../src/node_modules
+  export PATH=`pwd`/node/deps/npm/bin/node-gyp-bin/:`pwd`/node/out/Release:$PATH
+
+  banner "SPARK-WEBGL"
+
+  cd spark-webgl
+  node-gyp rebuild
+  cd ..
+
+fi
+#--------
 
 #--------  SQLITE
 
@@ -499,158 +532,169 @@ then
   cd ..
 fi
 
-if [[ $# -eq 1 ]] && [[ $1 == "SPARK_ENABLE_VIDEO" ]]; then
-#-------- cJSON
-
-if [ ! -e $EXT_INSTALL_PATH/lib/libcjson.$LIBEXTN ]
+if [[ $# -eq 1 ]] && [[ $1 == "SPARK_ENABLE_VIDEO" ]]
 then
-  banner "cJSON"
 
-  ./cJSON/build.sh
-fi
+  #-------- cJSON
+
+    if [ ! -e $EXT_INSTALL_PATH/lib/libcjson.$LIBEXTN ]
+    then
+      banner "cJSON"
+
+      ./cJSON/build.sh
+    fi
+
+  #--------
+
+  #-------- ORC
+
+  if [ ! -e $EXT_INSTALL_PATH/lib/liborc-0.4.la ]
+  then
+
+    banner "ORC"
+
+    ./orc/build.sh
+  fi
+
+  #--------
+
+  #-------- OSSP-UUID
+
+  if [ ! -e $EXT_INSTALL_PATH/lib/libuuid.la ]
+  then
+    banner "OSSP-UUID"
+
+    ./ossp-uuid/build.sh
+  fi
+
+  #--------
+
+  #--------libxml2
+
+  if [ ! -e $EXT_INSTALL_PATH/lib/libxml2.la ]
+  then
+    banner "libxml2"
+
+    ./libxml2/build.sh
+  fi
+
+  #--------
+
+  #-------- LIBDASH
+
+  if [ ! -e $EXT_INSTALL_PATH/lib/libdash.$LIBEXTN ]
+  then
+    banner "LIBDASH"
+
+    ./libdash/libdash/build.sh
+  fi
+
+  #--------
+
+  #-------- xz-5.2.2
+
+  if [ ! -e $EXT_INSTALL_PATH/lib/liblzma.la ]
+  then
+    banner "XZ"
+
+    ./xz/build.sh
+  fi
+
+  #--------
+
+  #-------- GSTREAMER-1.16
+
+  if [ ! -e $EXT_INSTALL_PATH/lib/libgstreamer-1.0.la ]
+  then
+
+    banner "GSTREAMER-1.16"
+
+    ./gstreamer/build.sh
+  fi
+
+  #--------
+
+  #-------- GST-PLUGIN-BASE
+
+  if [ ! -e $EXT_INSTALL_PATH/lib/libgstapp-1.0.la ]
+  then
+    banner "GST-PLUGINS-BASE"
+
+    ./gst-plugins-base/build.sh
+  fi
+
+  #--------
+
+  #-------- GST-PLUGIN-BAD
+
+  if [ ! -e $EXT_INSTALL_PATH/lib/libgstbadaudio-1.0.la ]
+  then
+    banner "GST-PLUGINS-BAD"
+
+    ./gst-plugins-bad/build.sh
+  fi
+
+  #--------
+
+  #-------- GST-PLUGIN-UGLY
+
+  if [ ! -e $EXT_INSTALL_PATH/lib/gstreamer-1.0/libgstx264.la ]
+  then
+    banner "GST-PLUGINS-UGLY"
+
+    ./gst-plugins-ugly/build.sh
+  fi
+
+  #--------
+
+  #-------- GST-PLUGIN-GOOD
+
+  if [ ! -e $EXT_INSTALL_PATH/lib/gstreamer-1.0/libgstavi.la ]
+  then
+    banner "GST-PLUGINS-GOOD"
+
+    ./gst-plugins-good/build.sh
+  fi
+
+  #--------
+
+  #-------- GST-LIBAV
+
+  if [ ! -e $EXT_INSTALL_PATH/lib/gstreamer-1.0/libgstlibav.la ]
+  then
+    banner "gst-libav"
+
+    ./gst-libav/build.sh
+  fi
+
+  #--------
+
+  #-------- AAMPABR
+
+  if [ ! -e $EXT_INSTALL_PATH/lib/libabr.$LIBEXTN ]
+  then
+    banner "AAMPABR"
+
+    ./aampabr/build.sh
+  fi
+
+  #--------
+
+  #-------- AAMP
+
+  if [ ! -e $EXT_INSTALL_PATH/lib/libaamp.$LIBEXTN ]
+  then
+    banner "AAMP"
+
+    ./aamp/build.sh
+  fi
 
 #--------
+fi # SPARK_ENABLE_VIDEO
 
-#--------orc
-
-if [ ! -e $EXT_INSTALL_PATH/lib/liborc-0.4.la ]
-then
-  banner "orc"
-
-  ./orc/build.sh
-fi
-
-#--------
-
-
-#--------ossp-uuid
-
-if [ ! -e $EXT_INSTALL_PATH/lib/libuuid.la ]
-then
-  banner "ossp-uuid"
-
-  ./ossp-uuid/build.sh
-fi
-
-#--------
-
-#--------libxml2
-
-if [ ! -e $EXT_INSTALL_PATH/lib/libxml2.la ]
-then
-  banner "libxml2"
-
-  ./libxml2/build.sh
-fi
-
-#--------
-
-#-------- libdash
-
-if [ ! -e $EXT_INSTALL_PATH/lib/libdash.$LIBEXTN ]
-then
-  banner "libdash"
-
-  ./libdash/libdash/build.sh
-fi
-
-#--------
-
-#-------- xz-5.2.2
-
-if [ ! -e $EXT_INSTALL_PATH/lib/liblzma.la ]
-then
-  banner "xz"
-
-  ./xz/build.sh
-fi
-
-#--------
-
-#-------- gstreamer-1.16
-
-if [ ! -e $EXT_INSTALL_PATH/lib/libgstreamer-1.0.la ]
-then
-  banner "gstreamer-1.16"
-
-  ./gstreamer/build.sh
-fi
-
-#--------
-
-#-------- gst-plugin-base
-
-if [ ! -e $EXT_INSTALL_PATH/lib/libgstapp-1.0.la ]
-then
-  banner "gst-plugins-base"
-
-  ./gst-plugins-base/build.sh
-fi
-
-#--------
-#-------- gst-plugin-bad
-
-if [ ! -e $EXT_INSTALL_PATH/lib/libgstbadaudio-1.0.la ]
-then
-  banner "gst-plugins-bad"
-
-  ./gst-plugins-bad/build.sh
-fi
-
-#--------
-#-------- gst-plugin-ugly
-
-if [ ! -e $EXT_INSTALL_PATH/lib/gstreamer-1.0/libgstx264.la ]
-then
-  banner "gst-plugins-ugly"
-
-  ./gst-plugins-ugly/build.sh
-fi
-
-#--------
-#-------- gst-plugin-good
-
-if [ ! -e $EXT_INSTALL_PATH/lib/gstreamer-1.0/libgstavi.la ]
-then
-  banner "gst-plugins-good"
-
-  ./gst-plugins-good/build.sh
-fi
-
-#--------
-#-------- gst-libav
-
-if [ ! -e $EXT_INSTALL_PATH/lib/gstreamer-1.0/libgstlibav.la ]
-then
-  banner "gst-libav"
-
-  ./gst-libav/build.sh
-fi
-
-#--------
-
-#-------- aampabr
-
-if [ ! -e $EXT_INSTALL_PATH/lib/libabr.$LIBEXTN ]
-then
-  banner "aampabr"
-
-  ./aampabr/build.sh
-fi
-
-#--------
-
-#-------- aamp
-
-if [ ! -e $EXT_INSTALL_PATH/lib/libaamp.$LIBEXTN ]
-then
-  banner "aamp"
-
-  ./aamp/build.sh
-fi
-
-#--------
-fi
+#--------------------------------------------
+banner ">>>>>  BUILD COMPLETE  <<<<<"
+#--------------------------------------------
 
 #-------
 exit 0    #success
